tools/requirements.txt
Pillow>=12.0.0
PyYAML>=6.0.1
pypdfium2>=5.4.0
